pyproject.toml
[tool.coverage.paths]
source = ["basicnanoclient", "*/site-packages"]

[tool.coverage.report]
show_missing = true

[build-system]
requires = ["setuptools", "wheel", "requests", "pynacl", "ed25519", "bitstring", "ed25519_blake2b"]

[project]
name = "basicnanoclient"
version = "0.1.3"
description = "Nano RPC python client"
urls = {'Bug Tracker' = 'https://github.com/nanoswap/basicnanoclient/issues'}
classifiers = [
    'Programming Language :: Python :: 3',
    'License :: OSI Approved :: The Unlicense (Unlicense)'
]
authors = [{'name' = 'Nathaniel Schultz', 'email' = 'nate@nanoswap.finance'}]
requires-python = '>=3.11'
readme = 'README.md'
license = {'file' = 'LICENSE'}
dependencies = [
    "requests",
    "pynacl",
    "ed25519_blake2b",
    "ed25519",
    "bitstring",
    "setuptools",
    "mkdocs",
    "mkdocstrings",
    "mkdocstrings-python"
]
